feat(config): reconciliation defaults and dust thresholds
Document the [reconciliation] config section in config.example.toml and
narrow the dust map's value type to u64 (the config crate's TOML
deserializer does not support u128, and u64 covers every Solana SPL
mint we quote). Adjusts the wallet and gateway monitors to widen u64
dust values into the u128 the drift window expects.

Add three config tests:
- config_default_reconciliation_uses_design_doc_values asserts the
  derived AppConfig::default reconciliation block matches the design
  doc (interval 10s, threshold 3, emit_event_on_skip true, dust
  USDC=10_000 / SOL=100_000 / cbBTC=100).
- config_example_reconciliation_section_matches_defaults parses the
  example file and asserts the same values.
- config_without_reconciliation_section_falls_back_to_defaults loads a
  minimal TOML with no [reconciliation] section and asserts every
  field falls back to its design-doc default.

+++ src/application/runtime/reconciliation/gateway_monitor.rs
@@ -50,7 +50,7 @@ impl GatewayMonitor {
         let dust_per_asset = config
             .dust
             .iter()
-            .map(|(asset, value)| (AssetId::from(asset.as_str()), *value))
+            .map(|(asset, value)| (AssetId::from(asset.as_str()), u128::from(*value)))
             .collect();
         Self {
             windows: HashMap::new(),
